# Optimize unnecessary check in Vec::retain