@TennyZhuang 2021-08-15
r? @oxalica  A trivial but interesting optimization, PTAL