@darksv 2021-08-16
@TennyZhuang I believe this change would break existing code relying on the fact that the passed closure is currently called once per item. 

From docs:
`This method operates in place, visiting each element exactly once in the original order, and preserves the order of the retained elements.`