@TennyZhuang 2021-08-16
> @TennyZhuang I believe this change would break existing code relying on the fact that the passed closure is currently called once per item.
> 
> From docs:
> `This method operates in place, visiting each element exactly once in the original order, and preserves the order of the retained elements.`

This change will not break it, each element is still called only once, we just split the loop to two stages, the processed_cnt will never go back.