@darksv 2021-08-16
> This change will not break it, each element is still operated exactly once in the original order, we just split the loop to two stages, the `processed_len` will never go back.

You're right. Sorry for interfering.