@TennyZhuang 2021-08-16
> I ran the benchmark on my R5 2400G PC and it is quite different, and is quite stable between multiple runs. The with-many-move case actually becomes slower for me.
> 
> Before
> 
> ```
> test vec::bench_retain_100000                            ... bench:      58,436 ns/iter (+/- 2,482)
> test vec::bench_retain_whole_100000                      ... bench:      54,747 ns/iter (+/- 1,646)
> ```
> 
> After
> 
> ```
> test vec::bench_retain_100000                            ... bench:      53,703 ns/iter (+/- 2,905)
> test vec::bench_retain_whole_100000                      ... bench:      61,588 ns/iter (+/- 689)
> ```
> 
> It might indicate that the generated code heavily relies on the _behavior_ of branch predictor. Are you on a Intel CPU? Intel's branch predictor may work different with AMD's.

I test it on my MacBook Pro 13 with `2.0GHz quad-core 10th-generation Intel Core i5 processor, Turbo Boost up to 3.8GHz`.

I can run the bench on another device.

@Xuanwo could you also run a benchmark for my optimization? 

@oxalica I will also try manually expand two loop without const generics.