@TennyZhuang 2021-08-16
@Xuanwo @oxalica Can you run a version without const generics?

For your convinience:

```
git remote add zty git@github.com:TennyZhuang/rust.git
git fetch zty
git checkout optimize-vec-retain-expand-manully
```