@TennyZhuang 2021-08-16
On my mac, the performance is same between const generics and manully expand:

```
test vec::bench_retain_100000                            ... bench:      52,471 ns/iter (+/- 3,921)
test vec::bench_retain_whole_100000                      ... bench:      38,560 ns/iter (+/- 2,666)
```