@TennyZhuang 2021-08-16
OK, the conclusion is that:

1. const generics is OK.
2. It optimize a little (10%) in most cases or architectures.
3. Performance is downgraded a little (10%) in some case and some specific architecture.
4. It optimize much (8x) on some specific architecture with unknown reason.
5. In theory, it actually optimize a unnecessary check.
6. It introduce a little code complexity.

I tend to think this PR deserves to be merged.

@oxalica what's your opinion?